chore: adjusting user-management

+++ app/Http/Controllers/Api/Admin/UserSettingsController.php
@@ -11,6 +11,7 @@ use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Validation\Rule;
 
 class UserSettingsController extends Controller
 {
@@ -60,21 +61,37 @@ class UserSettingsController extends Controller
         //
     }
 
+    public function edit($username) {
+        try {
+           $data = User::with('user_detail')->where('username', $username)->first();
+
+            Log::info('Berhasil mendapatkan data user');
+            return APIHelpers::responseAPI($data, 200);
+        } catch (Exception $error) {
+            Log::error('Gagal mendapatkan data user');
+            ActivityHelpers::LogActivityHelpers('Gagal mendapatkan data User! (Admin)', ['message' => $error->getMessage()], '0');
+            return APIHelpers::responseAPI([
+                'error' => $error->getMessage()
+            ], 500);
+        }
+    }
+
     /**
      * Store a newly created resource in storage.
      */
     public function storeOrUpdate(Request $request)
     {
         try {
+            $isUpdate = $request->has('username');
+
             $validate = $request->validate([
-                'username' => 'unique:users|required|string|min:5|max:20',
+                'username' => $isUpdate ? "" : 'unique:users|required|string|min:5|max:20',
                 'nama_depan' => 'required',
                 'nama_belakang' => 'nullable',
-                'email' => 'unique:users|email',
-                'password' => 'required|string',
+                'email' => $isUpdate ? '' : 'unique:users|email',
+                'password' => $isUpdate ? 'nullable|string' : 'required|string',
                 'avatar' => 'nullable|image|mimes:jpeg,png,jpg|max:2048',
                 'is_vet' => 'nullable|in:0,1',
-                'is_active' => 'nullable|in:0,1',
                 'alamat' => 'required|string',
                 'tempat_lahir' => 'required|string',
                 'tanggal_lahir' => 'required|date',
@@ -84,14 +101,20 @@ class UserSettingsController extends Controller
             ]);
     
             $data = [
-                'username' => $validate['username'],
                 'nama_depan' => $validate['nama_depan'],
                 'nama_belakang' => $validate['nama_belakang'],
                 'email' => $validate['email'],
-                'password' => Hash::make($validate['password']),
-                'is_active' => $validate['is_active'],
-                'is_vet' => $validate['is_vet'] ?? '0'
+                'is_active' => '1',
+                'is_vet' => (string) $validate['is_vet'] ?? '0'
             ];
+
+            if (!$isUpdate) {
+                $data['username'] = $validate['username'];
+            }
+
+            if ($request->password != '') {
+                $data['password'] = Hash::make($validate['password']);
+            }
     
             if ($request->hasFile('avatar')) {
                 $img = $validate['avatar']->store('user/avatar', 'public');
@@ -100,7 +123,7 @@ class UserSettingsController extends Controller
     
             User::updateOrCreate(['email' => $validate['email']], $data);
     
-            $user = User::where('username', $validate['username'])->first();
+            $user = User::where('email', $validate['email'])->first();
                 $userID = $user->id;
     
             $dataDetail = [
@@ -116,7 +139,7 @@ class UserSettingsController extends Controller
 
             Log::info('Berhasil Membuat Account! (Admin)');
             ActivityHelpers::LogActivityHelpers('Berhasil Membuat Account! (Admin)', ['user' => $data, 'detail' => $dataDetail], '1');
-            return APIHelpers::responseAPI(['message' => ['user' => $data, 'detail' => $dataDetail]], 200);
+            return APIHelpers::responseAPI(['message' => 'Berhasil Mengolah Data Account!', 'data' => ['user' => $data, 'detail' => $dataDetail]], 200);
         } catch (Exception $error) {
             Log::error($error->getMessage());
             ActivityHelpers::LogActivityHelpers('Gagal Membuat Account! (Admin)', ['message' => $error->getMessage()], '0');
@@ -125,15 +148,19 @@ class UserSettingsController extends Controller
 
     }
 
-    public function status($id) {
+    public function status($username) {
         try {
-            $data = User::with('user_detail')->findOrFail($id);
+            $data = User::with('user_detail')->where('username', $username)->first();
+            if (!$data) {
+                ActivityHelpers::LogActivityHelpers('Gagal Mengganti Status Account! (Admin)', ['message' => 'User tidak ditemukan'], '0');
+                return APIHelpers::responseAPI(['message' => 'User tidak ditemukan'], 404);
+            }
             $data->update([
-                'is_active' => $data->is_active == 0 ? 1 : 0
+                'is_active' => $data->is_active == 0 ? '1' : '0'
             ]);
             Log::info('Berhasil Mengganti Status Account! (Admin)');
             ActivityHelpers::LogActivityHelpers('Berhasil Mengganti Status Account! (Admin)', $data, '1');
-            return APIHelpers::responseAPI(['message' => $data], 200);
+            return APIHelpers::responseAPI(['message' => 'Berhasil Mengganti Status Account!', 'data' => $data], 200);
         } catch (Exception $error) {
             Log::error($error->getMessage());
             ActivityHelpers::LogActivityHelpers('Gagal Mengganti Status Account! (Admin)', ['message' => $error->getMessage()], '0');
@@ -141,20 +168,44 @@ class UserSettingsController extends Controller
         }
     }
 
+    public function vetStatus($username) {
+        try {
+            $data = User::with('user_detail')->where('username', $username)->first();
+            if (!$data) {
+                ActivityHelpers::LogActivityHelpers('Gagal Mengganti Status Account! (Admin)', ['message' => 'User tidak ditemukan'], '0');
+                return APIHelpers::responseAPI(['message' => 'User tidak ditemukan'], 404);
+            }
+            $data->update([
+                'is_vet' => $data->is_vet == 0 ? '1' : '0'
+            ]);
+            Log::info('Berhasil Mengganti Status Vet Account! (Admin)');
+            ActivityHelpers::LogActivityHelpers('Berhasil Mengganti Status Account! (Admin)', $data, '1');
+            return APIHelpers::responseAPI(['message' => 'Berhasil Mengganti Status Vet Account!', 'data' => $data], 200);
+        } catch (Exception $error) {
+            Log::error($error->getMessage());
+            ActivityHelpers::LogActivityHelpers('Gagal Mengganti Status Vet Account! (Admin)', ['message' => $error->getMessage()], '0');
+            return APIHelpers::responseAPI(['message' => $error->getMessage()], 500);
+        }
+    }
+
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $username)
     {
         try {
-            $data = User::with('user_detail')->findOrFail($id);
+            $data = User::with('user_detail')->where('username', $username)->first();
+            if (!$data) {
+                ActivityHelpers::LogActivityHelpers('Gagal Menghapus Account! (Admin)', ['message' => 'User tidak ditemukan'], '0');
+                return APIHelpers::responseAPI(['message' => 'User tidak ditemukan'], 404);
+            }
             $data->delete();
-            Log::info('Berhasil Menghapus Status Account! (Admin)');
-            ActivityHelpers::LogActivityHelpers('Berhasil Menghapus Status Account! (Admin)', $data, '1');
+            Log::info('Berhasil Menghapus Account! (Admin)');
+            ActivityHelpers::LogActivityHelpers('Berhasil Menghapus Account! (Admin)', $data, '1');
             return APIHelpers::responseAPI(['message' => $data], 200);
         } catch (Exception $error) {
             Log::error($error->getMessage());
-            ActivityHelpers::LogActivityHelpers('Gagal Menghapus Status Account! (Admin)', ['message' => $error->getMessage()], '0');
+            ActivityHelpers::LogActivityHelpers('Gagal Menghapus Account! (Admin)', ['message' => $error->getMessage()], '0');
             return APIHelpers::responseAPI(['message' => $error->getMessage()], 500);
         }
     }
